TheBlog/requirements.txt
asgiref==3.4.1
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0
django-mysql==4.3.0
django-rest-framework==0.1.0
django-social-auth==0.7.28
djangorestframework==3.13.1
factories==1.2.0
factory-boy==3.2.1
Faker==10.0.0
httplib2==0.20.2
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
oauth2==1.9.0.post1
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.0
pycparser==2.21
pycryptodome==3.12.0
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-openid==2.2.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
